nabu: aidl: hwcontrol: fix hal on android-14 QPR2.
Android 14 doesn't allow the running services to publish binder for hals that are not frozen.
it removes the hal manifest the manifest entry effectively rendering the hal service useless.

Without published binder interface XiaomiParts interation for pen,dt2w and keyboard code
can't connect with the service and issue commands to the underlying architecture to thereby
not allowing the peripherals to work

Refrence: https://source.android.com/docs/core/architecture/aidl/stable-aidl#manifests

As per the documentation this should have killed the hal on Android-15 but seems not,
even Android-14 QPR2 is affected with this change.

So, Let's stop being lazy and mark the interface for vintf stability and run the long
impending update + freeze api for hal and set owner as well while we are at it, also
update the Makefiles in parts and hal service code to use v1 of the interface.

Shout out to aosp on getting my lazy A$$ to work.

[1] https://cs.android.com/android/platform/superproject/main/+/main:system/tools/aidl/

-> After adding `@VintfStability` to aidl files run the below commands

`m custom.hardware.hwcontrol-update-api`
`m custom.hardware.hwcontrol-freeze-api`

Signed-off-by: Harshit Jain <reach@harsh1998.dev>